apply lua for fishingman, restructure some script files, add weather script to common
