+++ Assets/_Scripts/Controller/BookController/SwipeManager.cs
