+++ Assets/_Scripts/Controller/BookController/SwipeManager.cs.meta
@@ -1,0 +1,17 @@
+fileFormatVersion: 2
+<<<<<<< HEAD:Assets/_Scripts/Controller/Book2D/SwipeManager.cs.meta
+guid: 700577b74306143f1b268900591904f8
+timeCreated: 1480921641
+=======
+guid: 875d07e29bdc74a1d9c0588a6f48bc0a
+timeCreated: 1483090203
+>>>>>>> 80bb7a8f1f11d8771c341d3b90981d84b964ad6c:Assets/_Scripts/Controller/Fishingman/SwipeManager.cs.meta
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
