+++ Assets/_Scripts/Controller/BookController/WeatherController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: aedbe1747bbe644afba19dcaa0ec84eb
+timeCreated: 1483608679
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
